boot: move UEFI memory map structs to kernel-visible headers
We'll need the UEFI memory map in the kernel soon, so move the
data structures and helpers for interacting with the data passed
by loader(7) into a kernel-visble header, rather than restricting
visibility to dboot.
